refactor(mf): calendar - scss into css
