+++ micro-frontends/src/mf-calendar/src/ui/modules/days/days.module.css
@@ -1,0 +1,8 @@
+.days {
+	width: 160px;
+	height: 145px;
+	box-shadow: var(--ds-border-field);
+	padding: var(--ds-spacing-xs);
+	text-align: center;
+	background-color: var(--ds-background-color-secondary);
+}
