--- micro-frontends/src/mf-calendar/src/ui/modules/days/days.module.scss
@@ -1,8 +1,0 @@
-.days {
-  width: 160px;
-  height: 145px;
-  box-shadow: var(--border-field);
-  padding: 2px;
-  text-align: center;
-  background-color: var(--background-color-white);
-}
