--- micro-frontends/src/mf-calendar/src/ui/modules/days/month-days/month-days.module.scss
@@ -1,15 +1,0 @@
-.month-days {
-	display: grid;
-	grid-template-columns: repeat(7, 1fr);
-	grid-template-rows: repeat(6, 1fr);
-}
-
-.day {
-	height: 20px;
-	@include font('medium', 12, 20);
-}
-
-.day-active {
-	color: $cl-white;
-	background-color: $cl-blue000;
-}
