--- micro-frontends/src/mf-calendar/src/ui/modules/days/week-days/week-days.module.scss
@@ -1,12 +1,0 @@
-.week-days {
-	display: grid;
-	height: 20px;
-	grid-template-columns: repeat(7, 1fr);
-	background-color: $cl-grey80;
-	color: var(--text-color-3);
-}
-
-.day {
-	height: 20px;
-	@include font('medium', 12, 20);
-}
